Delete dead entity resolution code, keep graph-specific code
Delete entity_store.py and reconciliation.py (identity resolution now
owned by LML). Extract label CRUD to new label_store.py for
label_hierarchy.py. Create pipeline_db.py as the slimmed-down successor
to entity_store.py (schema setup, artist CRUD, stats, dedup only).

Delete 6 identity methods from wikidata_client.py (lookup_by_discogs_ids,
lookup_streaming_ids, search_by_name, search_musician_by_name,
search_musicians_batch, _filter_musicians) and their constants
(MUSICIAN_OCCUPATIONS, MUSICAL_GROUP_TYPES). Keep
lookup_labels_by_discogs_ids (used by label_hierarchy.py), get_influences,
get_label_hierarchy, and shared SPARQL infrastructure.

Delete 2 identity methods from musicbrainz_client.py (lookup_by_name,
batch_lookup). Keep get_recording_mbids for AcousticBrainz.

Remove --entity-store-path, --skip-reconciliation, --fetch-streaming-ids,
--entity-source, and --wikidata-reconciliation flags from run_pipeline.py.
Replace with --db-path flag. LML identity import is now the only identity
source path.

Remove ReconciliationEvent, ReconciliationReport, and WikidataStreamingIds
models. Update wikidata_influence.py to accept sqlite3.Connection instead
of EntityStore. Update sqlite_export.py to use PipelineDB. Update all
test fixtures and delete 9 test files for removed modules.
