--- semantic_index/entity_store.py
@@ -1,946 +1,0 @@
-"""Persistent entity store for reconciled artist identities.
-
-Manages the entity store tables within a SQLite database: entity (Wikidata QID,
-display name, type), artist migration (adds external ID columns to existing artist
-tables), and Phase 2 placeholder tables (release, label, label_hierarchy).
-
-Provides CRUD operations for entities, artist upsert with COALESCE semantics
-(never overwrites populated fields with NULL), and bulk stat updates.
-
-The ``initialize()`` method is fully idempotent — safe on a fresh database,
-an old-schema database, or an already-migrated database.
-"""
-
-from __future__ import annotations
-
-import logging
-import sqlite3
-from types import TracebackType
-
-from semantic_index.models import ArtistStats, DeduplicationReport, Entity, ReconciliationEvent
-from semantic_index.utils import ensure_columns
-
-logger = logging.getLogger(__name__)
-
-# Columns added to the artist table by _migrate_artist_table().
-# Each entry is (column_name, alter_table_definition).
-# ALTER TABLE ADD COLUMN requires constant defaults, so timestamps use NULL here
-# and are backfilled with strftime() after all columns are added.
-_NEW_ARTIST_COLUMNS = [
-    ("entity_id", "INTEGER REFERENCES entity(id)"),
-    ("musicbrainz_artist_id", "TEXT"),
-    ("wxyc_library_code_id", "INTEGER"),
-    ("reconciliation_status", "TEXT NOT NULL DEFAULT 'unreconciled'"),
-    ("created_at", "TEXT"),
-    ("updated_at", "TEXT"),
-]
-
-_ENTITY_STORE_SCHEMA = """
--- Real-world person, group, or organization.
-CREATE TABLE IF NOT EXISTS entity (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    wikidata_qid TEXT,
-    name TEXT NOT NULL,
-    entity_type TEXT NOT NULL DEFAULT 'artist',
-    spotify_artist_id TEXT,
-    apple_music_artist_id TEXT,
-    bandcamp_id TEXT,
-    created_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now')),
-    updated_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
-);
-
-CREATE INDEX IF NOT EXISTS idx_entity_qid ON entity(wikidata_qid) WHERE wikidata_qid IS NOT NULL;
-CREATE INDEX IF NOT EXISTS idx_entity_type ON entity(entity_type);
-
--- Phase 2 tables (created now, populated later)
-
-CREATE TABLE IF NOT EXISTS release (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    title TEXT NOT NULL,
-    artist_id INTEGER NOT NULL REFERENCES artist(id),
-    wxyc_library_release_id INTEGER,
-    discogs_master_id INTEGER,
-    discogs_release_id INTEGER,
-    musicbrainz_release_group_id TEXT,
-    year INTEGER,
-    created_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now')),
-    updated_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
-);
-
-CREATE TABLE IF NOT EXISTS label (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    name TEXT NOT NULL UNIQUE,
-    entity_id INTEGER REFERENCES entity(id),
-    discogs_label_id INTEGER,
-    musicbrainz_label_id TEXT,
-    created_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now')),
-    updated_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
-);
-
-CREATE TABLE IF NOT EXISTS label_hierarchy (
-    parent_label_id INTEGER NOT NULL REFERENCES label(id),
-    child_label_id INTEGER NOT NULL REFERENCES label(id),
-    source TEXT NOT NULL DEFAULT 'wikidata',
-    PRIMARY KEY (parent_label_id, child_label_id)
-);
-
--- Per-artist style tags from Discogs (persisted during reconciliation)
-
-CREATE TABLE IF NOT EXISTS artist_style (
-    artist_id INTEGER NOT NULL REFERENCES artist(id),
-    style_tag TEXT NOT NULL,
-    PRIMARY KEY (artist_id, style_tag)
-);
-
--- Reconciliation audit trail
-
-CREATE TABLE IF NOT EXISTS reconciliation_log (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    artist_id INTEGER NOT NULL REFERENCES artist(id),
-    source TEXT NOT NULL,
-    external_id TEXT NOT NULL,
-    confidence REAL,
-    method TEXT NOT NULL,
-    created_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
-);
-
-CREATE INDEX IF NOT EXISTS idx_reconciliation_artist ON reconciliation_log(artist_id);
-"""
-
-_ENTITY_INDEXES = """
-CREATE INDEX IF NOT EXISTS idx_entity_spotify ON entity(spotify_artist_id) WHERE spotify_artist_id IS NOT NULL;
-CREATE INDEX IF NOT EXISTS idx_entity_apple_music ON entity(apple_music_artist_id) WHERE apple_music_artist_id IS NOT NULL;
-CREATE INDEX IF NOT EXISTS idx_entity_bandcamp ON entity(bandcamp_id) WHERE bandcamp_id IS NOT NULL;
-"""
-
-_ARTIST_TABLE_SCHEMA = """
-CREATE TABLE IF NOT EXISTS artist (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    canonical_name TEXT NOT NULL UNIQUE,
-    genre TEXT,
-    total_plays INTEGER NOT NULL DEFAULT 0,
-    active_first_year INTEGER,
-    active_last_year INTEGER,
-    dj_count INTEGER NOT NULL DEFAULT 0,
-    request_ratio REAL NOT NULL DEFAULT 0.0,
-    show_count INTEGER NOT NULL DEFAULT 0,
-    discogs_artist_id INTEGER,
-    entity_id INTEGER REFERENCES entity(id),
-    musicbrainz_artist_id TEXT,
-    wxyc_library_code_id INTEGER,
-    reconciliation_status TEXT NOT NULL DEFAULT 'unreconciled',
-    created_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now')),
-    updated_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
-);
-"""
-
-_ARTIST_INDEXES = """
-CREATE INDEX IF NOT EXISTS idx_artist_entity ON artist(entity_id);
-CREATE INDEX IF NOT EXISTS idx_artist_discogs ON artist(discogs_artist_id) WHERE discogs_artist_id IS NOT NULL;
-CREATE INDEX IF NOT EXISTS idx_artist_musicbrainz ON artist(musicbrainz_artist_id) WHERE musicbrainz_artist_id IS NOT NULL;
-CREATE INDEX IF NOT EXISTS idx_artist_library_code ON artist(wxyc_library_code_id) WHERE wxyc_library_code_id IS NOT NULL;
-CREATE INDEX IF NOT EXISTS idx_artist_reconciliation ON artist(reconciliation_status);
-"""
-
-
-class EntityStore:
-    """Manages entity store tables within a SQLite database.
-
-    Args:
-        db_path: Path to the SQLite database file.
-    """
-
-    def __init__(self, db_path: str) -> None:
-        self._db_path = db_path
-        self._conn = sqlite3.connect(db_path)
-        self._conn.execute("PRAGMA journal_mode=WAL")
-        self._conn.execute("PRAGMA foreign_keys=ON")
-
-    def initialize(self) -> None:
-        """Create entity store tables and migrate the artist table if needed.
-
-        This is fully idempotent — safe to call on a fresh database, an
-        old-schema database, or an already-migrated database.
-        """
-        self._migrate_entity_unique_qid()
-        self._conn.executescript(_ENTITY_STORE_SCHEMA)
-        self._migrate_entity_table()
-        self._conn.executescript(_ENTITY_INDEXES)
-        self._migrate_artist_table()
-        if self._has_table("artist"):
-            self._conn.executescript(_ARTIST_INDEXES)
-        self._conn.commit()
-        logger.info("Entity store initialized: %s", self._db_path)
-
-    def _migrate_entity_unique_qid(self) -> None:
-        """Rebuild the entity table without the UNIQUE constraint on wikidata_qid.
-
-        Older databases have ``wikidata_qid TEXT UNIQUE``, which prevents
-        multiple entities from sharing a QID during reconciliation. This
-        migration detects the UNIQUE constraint via ``PRAGMA index_list``
-        and rebuilds the table without it so that the deduplication step
-        can find and merge duplicate-QID entities.
-
-        Safe to call when the entity table doesn't exist yet (no-op) or
-        when the constraint has already been removed (no-op).
-        """
-        if not self._has_table("entity"):
-            return
-
-        indexes = self._conn.execute("PRAGMA index_list(entity)").fetchall()
-        has_unique_qid = False
-        for idx in indexes:
-            idx_name = idx[1]
-            is_unique = bool(idx[2])
-            if is_unique:
-                cols = self._conn.execute(f"PRAGMA index_info('{idx_name}')").fetchall()
-                if any(col[2] == "wikidata_qid" for col in cols):
-                    has_unique_qid = True
-                    break
-
-        if not has_unique_qid:
-            return
-
-        logger.info("Rebuilding entity table to remove UNIQUE constraint on wikidata_qid")
-        self._conn.execute("PRAGMA foreign_keys=OFF")
-        self._conn.execute("ALTER TABLE entity RENAME TO _entity_old")
-        self._conn.execute(
-            """CREATE TABLE entity (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                wikidata_qid TEXT,
-                name TEXT NOT NULL,
-                entity_type TEXT NOT NULL DEFAULT 'artist',
-                created_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now')),
-                updated_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
-            )"""
-        )
-        self._conn.execute("INSERT INTO entity SELECT * FROM _entity_old")
-        self._conn.execute("DROP TABLE _entity_old")
-        self._conn.execute("PRAGMA foreign_keys=ON")
-        self._conn.commit()
-        logger.info("Entity table rebuilt without UNIQUE constraint on wikidata_qid")
-
-    # Columns to add to old entity tables that don't have them yet.
-    _NEW_ENTITY_COLUMNS: list[tuple[str, str]] = [
-        ("spotify_artist_id", "TEXT"),
-        ("apple_music_artist_id", "TEXT"),
-        ("bandcamp_id", "TEXT"),
-    ]
-
-    def _migrate_entity_table(self) -> None:
-        """Add new columns to an existing entity table if they are missing.
-
-        Uses the same PRAGMA table_info + ALTER TABLE ADD COLUMN pattern
-        as ``_migrate_artist_table()``. No-op if the entity table already
-        has all columns or does not exist yet (fresh-creation schema handles it).
-        """
-        if not self._has_table("entity"):
-            return
-
-        existing = {r[1] for r in self._conn.execute("PRAGMA table_info(entity)")}
-        for col_name, col_def in self._NEW_ENTITY_COLUMNS:
-            if col_name not in existing:
-                self._conn.execute(f"ALTER TABLE entity ADD COLUMN {col_name} {col_def}")
-                logger.info("Migrated entity table: added column %s", col_name)
-
-    def _has_table(self, name: str) -> bool:
-        """Check whether a table exists in the database."""
-        row = self._conn.execute(
-            "SELECT name FROM sqlite_master WHERE type='table' AND name=?", (name,)
-        ).fetchone()
-        return row is not None
-
-    def _migrate_artist_table(self) -> None:
-        """Create the artist table if missing, or add new columns to an existing one.
-
-        On a fresh database, creates the full artist table with all entity
-        store columns. On an old-schema database, uses PRAGMA table_info to
-        detect which columns already exist and only adds the missing ones.
-        """
-        if not self._has_table("artist"):
-            self._conn.executescript(_ARTIST_TABLE_SCHEMA)
-            logger.info("Created artist table with full entity store schema")
-            return
-
-        added = ensure_columns(self._conn, "artist", _NEW_ARTIST_COLUMNS)
-
-        # Backfill timestamps for existing rows (ALTER TABLE can't use strftime default)
-        if "created_at" in added or "updated_at" in added:
-            self._conn.execute(
-                """UPDATE artist
-                   SET created_at = COALESCE(created_at, strftime('%Y-%m-%dT%H:%M:%SZ', 'now')),
-                       updated_at = COALESCE(updated_at, strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))"""
-            )
-
-    # ------------------------------------------------------------------
-    # Entity CRUD
-    # ------------------------------------------------------------------
-
-    def get_or_create_entity(
-        self,
-        name: str,
-        entity_type: str,
-        wikidata_qid: str | None = None,
-    ) -> Entity:
-        """Return an existing entity by name+type, or create a new one.
-
-        If the entity already exists, its wikidata_qid is not overwritten.
-
-        Args:
-            name: Display name of the entity.
-            entity_type: One of 'artist', 'label', etc.
-            wikidata_qid: Optional Wikidata QID to set on creation.
-
-        Returns:
-            The existing or newly created Entity.
-        """
-        row = self._conn.execute(
-            "SELECT id, wikidata_qid, name, entity_type FROM entity WHERE name = ? AND entity_type = ?",
-            (name, entity_type),
-        ).fetchone()
-        if row is not None:
-            return Entity(id=row[0], wikidata_qid=row[1], name=row[2], entity_type=row[3])
-
-        cur = self._conn.execute(
-            "INSERT INTO entity (name, entity_type, wikidata_qid) VALUES (?, ?, ?)",
-            (name, entity_type, wikidata_qid),
-        )
-        self._conn.commit()
-        return Entity(
-            id=cur.lastrowid,  # type: ignore[arg-type]
-            name=name,
-            entity_type=entity_type,
-            wikidata_qid=wikidata_qid,
-        )
-
-    def update_entity_qid(self, entity_id: int, wikidata_qid: str) -> None:
-        """Set the Wikidata QID for an entity.
-
-        Args:
-            entity_id: The entity's primary key.
-            wikidata_qid: The Wikidata QID to assign.
-
-        Raises:
-            ValueError: If no entity with the given id exists.
-        """
-        cur = self._conn.execute(
-            "UPDATE entity SET wikidata_qid = ?, updated_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now') WHERE id = ?",
-            (wikidata_qid, entity_id),
-        )
-        if cur.rowcount == 0:
-            raise ValueError(f"No entity with id {entity_id}")
-        self._conn.commit()
-
-    def get_entity_by_qid(self, wikidata_qid: str) -> Entity | None:
-        """Look up an entity by Wikidata QID.
-
-        Args:
-            wikidata_qid: The Wikidata QID to search for.
-
-        Returns:
-            The matching Entity, or None if not found.
-        """
-        row = self._conn.execute(
-            "SELECT id, wikidata_qid, name, entity_type FROM entity WHERE wikidata_qid = ?",
-            (wikidata_qid,),
-        ).fetchone()
-        if row is None:
-            return None
-        return Entity(id=row[0], wikidata_qid=row[1], name=row[2], entity_type=row[3])
-
-    def merge_entities(self, keep_id: int, merge_id: int) -> None:
-        """Merge two entities: re-parent artists from merge_id to keep_id, then delete merge_id.
-
-        Args:
-            keep_id: The entity ID to keep.
-            merge_id: The entity ID to merge into keep_id and delete.
-
-        Raises:
-            ValueError: If keep_id == merge_id, or if either entity doesn't exist.
-        """
-        if keep_id == merge_id:
-            raise ValueError("Cannot merge an entity into itself")
-
-        for eid, label in [(keep_id, "keep"), (merge_id, "merge")]:
-            row = self._conn.execute("SELECT id FROM entity WHERE id = ?", (eid,)).fetchone()
-            if row is None:
-                raise ValueError(f"No entity with id {eid} ({label})")
-
-        self._conn.execute(
-            "UPDATE artist SET entity_id = ? WHERE entity_id = ?", (keep_id, merge_id)
-        )
-        self._conn.execute("DELETE FROM entity WHERE id = ?", (merge_id,))
-        self._conn.commit()
-
-    # ------------------------------------------------------------------
-    # Entity Deduplication
-    # ------------------------------------------------------------------
-
-    def find_duplicate_qid_groups(self) -> list[tuple[str, list[int]]]:
-        """Find groups of entities sharing the same non-NULL Wikidata QID.
-
-        Returns:
-            List of ``(wikidata_qid, [entity_id, ...])`` tuples for QIDs
-            with two or more entities. Entity IDs are sorted ascending.
-        """
-        rows = self._conn.execute(
-            """SELECT wikidata_qid, GROUP_CONCAT(id)
-               FROM entity
-               WHERE wikidata_qid IS NOT NULL
-               GROUP BY wikidata_qid
-               HAVING COUNT(*) > 1
-               ORDER BY wikidata_qid"""
-        ).fetchall()
-        return [(row[0], sorted(int(x) for x in row[1].split(","))) for row in rows]
-
-    def deduplicate_by_qid(self) -> DeduplicationReport:
-        """Find entities sharing a Wikidata QID and merge duplicates.
-
-        For each QID with multiple entities, keeps the entity with the
-        lowest ID and merges the rest into it using ``merge_entities()``.
-
-        Returns:
-            DeduplicationReport with counts of groups found, entities
-            merged, and artists reassigned.
-        """
-        groups = self.find_duplicate_qid_groups()
-        entities_merged = 0
-        artists_reassigned = 0
-
-        for qid, entity_ids in groups:
-            keep_id = entity_ids[0]
-            for merge_id in entity_ids[1:]:
-                count = self._conn.execute(
-                    "SELECT COUNT(*) FROM artist WHERE entity_id = ?", (merge_id,)
-                ).fetchone()[0]
-                artists_reassigned += count
-                self.merge_entities(keep_id, merge_id)
-                entities_merged += 1
-            logger.info(
-                "Deduplicated QID %s: kept entity %d, merged %d entities",
-                qid,
-                keep_id,
-                len(entity_ids) - 1,
-            )
-
-        if groups:
-            logger.info(
-                "Entity deduplication: %d groups, %d entities merged, %d artists reassigned",
-                len(groups),
-                entities_merged,
-                artists_reassigned,
-            )
-        else:
-            logger.info("Entity deduplication: no duplicate QIDs found")
-
-        return DeduplicationReport(
-            groups_found=len(groups),
-            entities_merged=entities_merged,
-            artists_reassigned=artists_reassigned,
-        )
-
-    # ------------------------------------------------------------------
-    # Artist Upsert
-    # ------------------------------------------------------------------
-
-    def upsert_artist(
-        self,
-        canonical_name: str,
-        *,
-        genre: str | None = None,
-        discogs_artist_id: int | None = None,
-        entity_id: int | None = None,
-        musicbrainz_artist_id: str | None = None,
-        wxyc_library_code_id: int | None = None,
-    ) -> int:
-        """Insert or update an artist row using COALESCE semantics.
-
-        On conflict (canonical_name), only updates fields where the new value
-        is not NULL — existing populated fields are never overwritten with NULL.
-
-        Args:
-            canonical_name: The artist's canonical name (unique key).
-            genre: Genre string.
-            discogs_artist_id: Discogs artist ID.
-            entity_id: FK to the entity table.
-            musicbrainz_artist_id: MusicBrainz artist UUID.
-            wxyc_library_code_id: WXYC library code ID.
-
-        Returns:
-            The artist row's integer primary key.
-        """
-        cur = self._conn.execute(
-            """INSERT INTO artist (canonical_name, genre, discogs_artist_id, entity_id,
-                                   musicbrainz_artist_id, wxyc_library_code_id,
-                                   updated_at)
-               VALUES (?, ?, ?, ?, ?, ?, strftime('%Y-%m-%dT%H:%M:%SZ', 'now'))
-               ON CONFLICT(canonical_name) DO UPDATE SET
-                   genre = COALESCE(excluded.genre, artist.genre),
-                   discogs_artist_id = COALESCE(excluded.discogs_artist_id, artist.discogs_artist_id),
-                   entity_id = COALESCE(excluded.entity_id, artist.entity_id),
-                   musicbrainz_artist_id = COALESCE(excluded.musicbrainz_artist_id, artist.musicbrainz_artist_id),
-                   wxyc_library_code_id = COALESCE(excluded.wxyc_library_code_id, artist.wxyc_library_code_id),
-                   updated_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-               RETURNING id""",
-            (
-                canonical_name,
-                genre,
-                discogs_artist_id,
-                entity_id,
-                musicbrainz_artist_id,
-                wxyc_library_code_id,
-            ),
-        )
-        row = cur.fetchone()
-        self._conn.commit()
-        assert row is not None  # RETURNING always yields a row
-        return int(row[0])
-
-    def get_artist_by_name(self, canonical_name: str) -> dict[str, object] | None:
-        """Look up an artist row by canonical name.
-
-        Args:
-            canonical_name: The artist's canonical name.
-
-        Returns:
-            A dict of column name -> value, or None if not found.
-        """
-        self._conn.row_factory = sqlite3.Row
-        row = self._conn.execute(
-            "SELECT * FROM artist WHERE canonical_name = ?", (canonical_name,)
-        ).fetchone()
-        self._conn.row_factory = None
-        if row is None:
-            return None
-        return dict(row)
-
-    def bulk_upsert_artists(self, names: list[str]) -> dict[str, int]:
-        """Insert or retrieve artist rows for a list of canonical names.
-
-        This is the main pipeline entry point for ensuring artist rows exist.
-        Deduplicates the input list.
-
-        Args:
-            names: List of canonical artist names.
-
-        Returns:
-            Dict mapping canonical_name -> artist row id.
-        """
-        unique_names = list(dict.fromkeys(names))
-        result: dict[str, int] = {}
-        for name in unique_names:
-            result[name] = self.upsert_artist(name)
-        return result
-
-    # ------------------------------------------------------------------
-    # Reconciliation Log
-    # ------------------------------------------------------------------
-
-    def log_reconciliation(
-        self,
-        artist_id: int,
-        source: str,
-        external_id: str,
-        confidence: float | None,
-        method: str,
-    ) -> None:
-        """Record a reconciliation event for an artist.
-
-        Args:
-            artist_id: The artist row's primary key.
-            source: External knowledge base ('discogs', 'musicbrainz', 'wikidata').
-            external_id: Identifier in the external source.
-            confidence: Match confidence score, or None if not applicable.
-            method: Matching method ('exact', 'fuzzy', 'api_search', 'cache_lookup').
-
-        Raises:
-            ValueError: If no artist with the given id exists.
-        """
-        row = self._conn.execute("SELECT id FROM artist WHERE id = ?", (artist_id,)).fetchone()
-        if row is None:
-            raise ValueError(f"No artist with id {artist_id}")
-        self._conn.execute(
-            """INSERT INTO reconciliation_log (artist_id, source, external_id, confidence, method)
-               VALUES (?, ?, ?, ?, ?)""",
-            (artist_id, source, external_id, confidence, method),
-        )
-        self._conn.commit()
-
-    def get_reconciliation_history(self, artist_id: int) -> list[ReconciliationEvent]:
-        """Return all reconciliation events for an artist, ordered by insertion.
-
-        Args:
-            artist_id: The artist row's primary key.
-
-        Returns:
-            List of ReconciliationEvent instances, oldest first.
-        """
-        rows = self._conn.execute(
-            """SELECT source, external_id, confidence, method
-               FROM reconciliation_log
-               WHERE artist_id = ?
-               ORDER BY id""",
-            (artist_id,),
-        ).fetchall()
-        return [
-            ReconciliationEvent(
-                source=row[0],
-                external_id=row[1],
-                confidence=row[2],
-                method=row[3],
-            )
-            for row in rows
-        ]
-
-    # ------------------------------------------------------------------
-    # Reconciliation Queries
-    # ------------------------------------------------------------------
-
-    def get_unreconciled_artists(self, limit: int | None = None) -> list[tuple[int, str]]:
-        """Return (id, canonical_name) pairs for artists with status 'unreconciled'.
-
-        Args:
-            limit: Maximum number of artists to return. None for all.
-
-        Returns:
-            List of (artist_id, canonical_name) tuples.
-        """
-        sql = "SELECT id, canonical_name FROM artist WHERE reconciliation_status = 'unreconciled'"
-        if limit is not None:
-            sql += f" LIMIT {limit}"
-        rows = self._conn.execute(sql).fetchall()
-        return [(row[0], row[1]) for row in rows]
-
-    def get_no_match_artists(self, limit: int | None = None) -> list[tuple[int, str]]:
-        """Return (id, canonical_name) pairs for artists with status 'no_match'.
-
-        Args:
-            limit: Maximum number of artists to return. None for all.
-
-        Returns:
-            List of (artist_id, canonical_name) tuples.
-        """
-        sql = "SELECT id, canonical_name FROM artist WHERE reconciliation_status = 'no_match'"
-        if limit is not None:
-            sql += f" LIMIT {limit}"
-        rows = self._conn.execute(sql).fetchall()
-        return [(row[0], row[1]) for row in rows]
-
-    def get_artists_needing_wikidata(self) -> list[tuple[int, str, int]]:
-        """Return artists that have a Discogs ID but no Wikidata QID yet.
-
-        An artist "needs Wikidata" when it has a ``discogs_artist_id`` AND
-        either has no linked entity or its linked entity has a NULL
-        ``wikidata_qid``.
-
-        Returns:
-            List of (artist_id, canonical_name, discogs_artist_id) tuples.
-        """
-        rows = self._conn.execute(
-            """SELECT a.id, a.canonical_name, a.discogs_artist_id
-               FROM artist a
-               LEFT JOIN entity e ON a.entity_id = e.id
-               WHERE a.discogs_artist_id IS NOT NULL
-                 AND (a.entity_id IS NULL OR e.wikidata_qid IS NULL)"""
-        ).fetchall()
-        return [(row[0], row[1], row[2]) for row in rows]
-
-    def get_entities_needing_streaming_ids(self) -> list[tuple[int, str]]:
-        """Return entities that have a Wikidata QID but no streaming IDs yet.
-
-        An entity "needs streaming IDs" when it has a non-NULL
-        ``wikidata_qid`` AND all three streaming ID columns are NULL.
-
-        Returns:
-            List of (entity_id, wikidata_qid) tuples.
-        """
-        rows = self._conn.execute(
-            """SELECT id, wikidata_qid FROM entity
-               WHERE wikidata_qid IS NOT NULL
-                 AND spotify_artist_id IS NULL
-                 AND apple_music_artist_id IS NULL
-                 AND bandcamp_id IS NULL"""
-        ).fetchall()
-        return [(row[0], row[1]) for row in rows]
-
-    def update_entity_streaming_ids(
-        self,
-        entity_id: int,
-        spotify: str | None = None,
-        apple_music: str | None = None,
-        bandcamp: str | None = None,
-    ) -> None:
-        """Update streaming service IDs for an entity using COALESCE semantics.
-
-        Existing non-NULL values are never overwritten with NULL. Only
-        non-NULL arguments replace current values.
-
-        Args:
-            entity_id: The entity row's primary key.
-            spotify: Spotify artist ID (P1902).
-            apple_music: Apple Music artist ID (P2850).
-            bandcamp: Bandcamp profile ID (P3283).
-        """
-        self._conn.execute(
-            """UPDATE entity
-               SET spotify_artist_id = COALESCE(?, spotify_artist_id),
-                   apple_music_artist_id = COALESCE(?, apple_music_artist_id),
-                   bandcamp_id = COALESCE(?, bandcamp_id),
-                   updated_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-               WHERE id = ?""",
-            (spotify, apple_music, bandcamp, entity_id),
-        )
-        self._conn.commit()
-
-    def update_reconciliation_status(self, artist_id: int, status: str) -> None:
-        """Update the reconciliation status for an artist.
-
-        Args:
-            artist_id: The artist row's primary key.
-            status: New status ('reconciled', 'no_match', 'partial', etc.).
-
-        Raises:
-            ValueError: If no artist with the given id exists.
-        """
-        cur = self._conn.execute(
-            """UPDATE artist SET reconciliation_status = ?,
-                   updated_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-               WHERE id = ?""",
-            (status, artist_id),
-        )
-        if cur.rowcount == 0:
-            raise ValueError(f"No artist with id {artist_id}")
-        self._conn.commit()
-
-    # ------------------------------------------------------------------
-    # Artist Styles
-    # ------------------------------------------------------------------
-
-    def persist_artist_styles(self, artist_id: int, styles: list[str]) -> None:
-        """Persist Discogs style tags for an artist (idempotent).
-
-        Uses INSERT OR IGNORE so overlapping styles from repeat calls
-        do not raise errors or create duplicates.
-
-        Args:
-            artist_id: The artist row's primary key.
-            styles: List of Discogs style strings.
-        """
-        if not styles:
-            return
-        self._conn.executemany(
-            "INSERT OR IGNORE INTO artist_style (artist_id, style_tag) VALUES (?, ?)",
-            [(artist_id, style) for style in styles],
-        )
-        self._conn.commit()
-
-    def get_artist_styles(self, artist_id: int) -> list[str]:
-        """Return all style tags for an artist, sorted alphabetically.
-
-        Args:
-            artist_id: The artist row's primary key.
-
-        Returns:
-            Sorted list of style strings.
-        """
-        rows = self._conn.execute(
-            "SELECT style_tag FROM artist_style WHERE artist_id = ? ORDER BY style_tag",
-            (artist_id,),
-        ).fetchall()
-        return [row[0] for row in rows]
-
-    # ------------------------------------------------------------------
-    # Label CRUD
-    # ------------------------------------------------------------------
-
-    def get_or_create_label(
-        self,
-        name: str,
-        *,
-        discogs_label_id: int | None = None,
-        wikidata_qid: str | None = None,
-    ) -> int:
-        """Return an existing label by name, or create a new one.
-
-        On conflict (name), existing discogs_label_id is not overwritten.
-        If a wikidata_qid is provided, an entity row is also created/linked.
-
-        Args:
-            name: Label name (unique key).
-            discogs_label_id: Optional Discogs label ID.
-            wikidata_qid: Optional Wikidata QID. Creates an entity if provided.
-
-        Returns:
-            The label row's integer primary key.
-        """
-        row = self._conn.execute("SELECT id FROM label WHERE name = ?", (name,)).fetchone()
-        if row is not None:
-            return int(row[0])
-
-        entity_id = None
-        if wikidata_qid:
-            entity = self.get_or_create_entity(name, "label", wikidata_qid=wikidata_qid)
-            entity_id = entity.id
-
-        cur = self._conn.execute(
-            "INSERT INTO label (name, discogs_label_id, entity_id) VALUES (?, ?, ?)",
-            (name, discogs_label_id, entity_id),
-        )
-        self._conn.commit()
-        return int(cur.lastrowid)  # type: ignore[arg-type]
-
-    def update_label_qid(self, label_id: int, wikidata_qid: str) -> None:
-        """Set the Wikidata QID for a label by creating/linking an entity.
-
-        Args:
-            label_id: The label's primary key.
-            wikidata_qid: The Wikidata QID to assign.
-
-        Raises:
-            ValueError: If no label with the given id exists.
-        """
-        row = self._conn.execute(
-            "SELECT id, name, entity_id FROM label WHERE id = ?", (label_id,)
-        ).fetchone()
-        if row is None:
-            raise ValueError(f"No label with id {label_id}")
-
-        label_name = row[1]
-        existing_entity_id = row[2]
-
-        if existing_entity_id is not None:
-            # Update the existing entity's QID
-            self._conn.execute(
-                "UPDATE entity SET wikidata_qid = COALESCE(wikidata_qid, ?), "
-                "updated_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now') WHERE id = ?",
-                (wikidata_qid, existing_entity_id),
-            )
-        else:
-            entity = self.get_or_create_entity(label_name, "label", wikidata_qid=wikidata_qid)
-            self._conn.execute(
-                "UPDATE label SET entity_id = ?, updated_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now') WHERE id = ?",
-                (entity.id, label_id),
-            )
-        self._conn.commit()
-
-    def insert_label_hierarchy(
-        self, parent_label_id: int, child_label_id: int, source: str = "wikidata"
-    ) -> None:
-        """Insert a parent-child label relationship (idempotent).
-
-        Args:
-            parent_label_id: The parent label's primary key.
-            child_label_id: The child label's primary key.
-            source: Source of the relationship (default: 'wikidata').
-        """
-        self._conn.execute(
-            "INSERT OR IGNORE INTO label_hierarchy (parent_label_id, child_label_id, source) VALUES (?, ?, ?)",
-            (parent_label_id, child_label_id, source),
-        )
-        self._conn.commit()
-
-    def get_labels_with_discogs_id(self) -> list[tuple[int, str, int]]:
-        """Return all labels that have a Discogs label ID.
-
-        Returns:
-            List of (label_id, name, discogs_label_id) tuples.
-        """
-        rows = self._conn.execute(
-            "SELECT id, name, discogs_label_id FROM label WHERE discogs_label_id IS NOT NULL"
-        ).fetchall()
-        return [(row[0], row[1], row[2]) for row in rows]
-
-    def get_label_by_name(self, name: str) -> dict[str, object] | None:
-        """Look up a label row by name.
-
-        Args:
-            name: The label name.
-
-        Returns:
-            A dict of column name -> value, or None if not found.
-        """
-        self._conn.row_factory = sqlite3.Row
-        row = self._conn.execute("SELECT * FROM label WHERE name = ?", (name,)).fetchone()
-        self._conn.row_factory = None
-        if row is None:
-            return None
-        return dict(row)
-
-    # ------------------------------------------------------------------
-    # Name-to-ID Mapping
-    # ------------------------------------------------------------------
-
-    def get_name_to_id_mapping(self) -> dict[str, int]:
-        """Return a mapping of canonical_name to artist row id.
-
-        This replaces the inline dict-building pattern in sqlite_export.py.
-
-        Returns:
-            Dict mapping canonical_name -> artist row id.
-        """
-        rows = self._conn.execute("SELECT id, canonical_name FROM artist").fetchall()
-        return {row[1]: row[0] for row in rows}
-
-    # ------------------------------------------------------------------
-    # Artist Stats
-    # ------------------------------------------------------------------
-
-    def update_artist_stats(self, canonical_name: str, stats: ArtistStats) -> None:
-        """Update stats columns for a single artist.
-
-        Args:
-            canonical_name: The artist's canonical name.
-            stats: The ArtistStats to apply.
-
-        Raises:
-            ValueError: If no artist with the given name exists.
-        """
-        cur = self._conn.execute(
-            """UPDATE artist SET
-                   total_plays = ?,
-                   genre = COALESCE(?, genre),
-                   active_first_year = ?,
-                   active_last_year = ?,
-                   dj_count = ?,
-                   request_ratio = ?,
-                   show_count = ?,
-                   updated_at = strftime('%Y-%m-%dT%H:%M:%SZ', 'now')
-               WHERE canonical_name = ?""",
-            (
-                stats.total_plays,
-                stats.genre,
-                stats.active_first_year,
-                stats.active_last_year,
-                stats.dj_count,
-                stats.request_ratio,
-                stats.show_count,
-                canonical_name,
-            ),
-        )
-        if cur.rowcount == 0:
-            raise ValueError(f"No artist with canonical_name '{canonical_name}'")
-        self._conn.commit()
-
-    def bulk_update_stats(self, artist_stats: dict[str, ArtistStats]) -> None:
-        """Update stats for multiple artists in a single transaction.
-
-        Args:
-            artist_stats: Dict mapping canonical_name -> ArtistStats.
-        """
-        for name, stats in artist_stats.items():
-            self.update_artist_stats(name, stats)
-
-    def close(self) -> None:
-        """Close the database connection."""
-        self._conn.close()
-
-    def __enter__(self) -> EntityStore:
-        return self
-
-    def __exit__(
-        self,
-        exc_type: type[BaseException] | None,
-        exc_val: BaseException | None,
-        exc_tb: TracebackType | None,
-    ) -> None:
-        self.close()
