+++ semantic_index/label_hierarchy.py
@@ -14,7 +14,7 @@ from typing import TYPE_CHECKING
 from semantic_index.models import LabelHierarchyReport
 
 if TYPE_CHECKING:
-    from semantic_index.entity_store import EntityStore
+    from semantic_index.label_store import LabelStore
     from semantic_index.models import ArtistEnrichment
     from semantic_index.wikidata_client import WikidataClient
 
@@ -22,7 +22,7 @@ logger = logging.getLogger(__name__)
 
 
 def populate_label_hierarchy(
-    entity_store: EntityStore,
+    label_store: LabelStore,
     enrichments: dict[str, ArtistEnrichment],
     wikidata_client: WikidataClient,
 ) -> LabelHierarchyReport:
@@ -37,7 +37,7 @@ def populate_label_hierarchy(
         6. Insert hierarchy edges (creating parent/child labels as needed).
 
     Args:
-        entity_store: The EntityStore managing the SQLite database.
+        label_store: The LabelStore managing the SQLite database.
         enrichments: Discogs enrichment data keyed by canonical artist name.
         wikidata_client: WikidataClient for SPARQL lookups.
 
@@ -62,7 +62,7 @@ def populate_label_hierarchy(
     # Step 2: Insert labels into entity store
     label_name_to_id: dict[str, int] = {}
     for name, discogs_id in unique_labels.items():
-        label_name_to_id[name] = entity_store.get_or_create_label(name, discogs_label_id=discogs_id)
+        label_name_to_id[name] = label_store.get_or_create_label(name, discogs_label_id=discogs_id)
 
     labels_created = len(label_name_to_id)
 
@@ -84,7 +84,7 @@ def populate_label_hierarchy(
         if label_name is None:
             continue
         label_id = label_name_to_id[label_name]
-        entity_store.update_label_qid(label_id, wd_entity.qid)
+        label_store.update_label_qid(label_id, wd_entity.qid)
         qid_to_label_name[wd_entity.qid] = label_name
 
     labels_matched = len(qid_map)
@@ -108,22 +108,22 @@ def populate_label_hierarchy(
         # Ensure parent label exists
         parent_name = rel.parent_name
         if parent_name not in label_name_to_id:
-            label_name_to_id[parent_name] = entity_store.get_or_create_label(
+            label_name_to_id[parent_name] = label_store.get_or_create_label(
                 parent_name, wikidata_qid=rel.parent_qid
             )
         else:
-            entity_store.update_label_qid(label_name_to_id[parent_name], rel.parent_qid)
+            label_store.update_label_qid(label_name_to_id[parent_name], rel.parent_qid)
 
         # Ensure child label exists
         child_name = rel.child_name
         if child_name not in label_name_to_id:
-            label_name_to_id[child_name] = entity_store.get_or_create_label(
+            label_name_to_id[child_name] = label_store.get_or_create_label(
                 child_name, wikidata_qid=rel.child_qid
             )
         else:
-            entity_store.update_label_qid(label_name_to_id[child_name], rel.child_qid)
+            label_store.update_label_qid(label_name_to_id[child_name], rel.child_qid)
 
-        entity_store.insert_label_hierarchy(
+        label_store.insert_label_hierarchy(
             label_name_to_id[parent_name],
             label_name_to_id[child_name],
         )
