+++ tests/conftest.py
@@ -13,7 +13,6 @@ from semantic_index.models import (
     DiscogsLabel,
     DiscogsRelease,
     DiscogsTrack,
-    Entity,
     FlowsheetEntry,
     LabelInfo,
     LibraryCode,
@@ -163,18 +162,6 @@ def make_personnel_credit(
         name=name,
         roles=roles if roles is not None else ["Written-By"],
     )
-
-
-def make_entity(
-    name: str = "Autechre",
-    entity_type: str = "artist",
-    wikidata_qid: str | None = None,
-) -> Entity:
-    """Create an Entity instance for testing.
-
-    Uses a fixed id=1 since the real id is assigned by SQLite AUTOINCREMENT.
-    """
-    return Entity(id=1, name=name, entity_type=entity_type, wikidata_qid=wikidata_qid)
 
 
 def make_wikidata_entity(
