+++ tests/unit/test_label_hierarchy.py
@@ -5,16 +5,17 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from semantic_index.entity_store import EntityStore
 from semantic_index.label_hierarchy import populate_label_hierarchy
+from semantic_index.label_store import LabelStore
 from semantic_index.models import (
     ArtistEnrichment,
     LabelInfo,
     WikidataEntity,
     WikidataLabelHierarchy,
 )
+from semantic_index.pipeline_db import PipelineDB
 
-# Old artist schema for EntityStore migration
+# Old artist schema for PipelineDB migration
 _OLD_ARTIST_SCHEMA = """
 CREATE TABLE artist (
     id INTEGER PRIMARY KEY AUTOINCREMENT,
@@ -32,14 +33,34 @@ CREATE TABLE artist (
 
 
 @pytest.fixture()
-def store(tmp_path) -> EntityStore:
+def label_store(tmp_path) -> LabelStore:
     db_path = str(tmp_path / "test.db")
     conn = sqlite3.connect(db_path)
     conn.executescript(_OLD_ARTIST_SCHEMA)
     conn.close()
-    s = EntityStore(db_path)
-    s.initialize()
-    return s
+    db = PipelineDB(db_path)
+    db.initialize()
+    return LabelStore(db._conn)
+
+
+def _get_label_by_name(conn: sqlite3.Connection, name: str) -> dict | None:
+    """Look up a label row by name."""
+    conn.row_factory = sqlite3.Row
+    row = conn.execute("SELECT * FROM label WHERE name = ?", (name,)).fetchone()
+    conn.row_factory = None
+    if row is None:
+        return None
+    return dict(row)
+
+
+def _get_entity_by_qid(conn: sqlite3.Connection, qid: str) -> dict | None:
+    """Look up an entity by Wikidata QID."""
+    conn.row_factory = sqlite3.Row
+    row = conn.execute("SELECT * FROM entity WHERE wikidata_qid = ?", (qid,)).fetchone()
+    conn.row_factory = None
+    if row is None:
+        return None
+    return dict(row)
 
 
 def _make_enrichments(
@@ -56,7 +77,7 @@ def _make_enrichments(
 
 
 class TestPopulateLabelHierarchy:
-    def test_creates_labels_from_enrichments(self, store: EntityStore):
+    def test_creates_labels_from_enrichments(self, label_store: LabelStore):
         enrichments = _make_enrichments(
             {
                 "Autechre": [("Warp Records", 23528)],
@@ -67,17 +88,17 @@ class TestPopulateLabelHierarchy:
         wikidata_client.lookup_labels_by_discogs_ids.return_value = {}
         wikidata_client.get_label_hierarchy.return_value = []
 
-        populate_label_hierarchy(store, enrichments, wikidata_client)
+        populate_label_hierarchy(label_store, enrichments, wikidata_client)
 
-        warp = store.get_label_by_name("Warp Records")
+        warp = _get_label_by_name(label_store._conn, "Warp Records")
         assert warp is not None
         assert warp["discogs_label_id"] == 23528
 
-        subpop = store.get_label_by_name("Sub Pop")
+        subpop = _get_label_by_name(label_store._conn, "Sub Pop")
         assert subpop is not None
         assert subpop["discogs_label_id"] == 1594
 
-    def test_links_wikidata_qids_to_labels(self, store: EntityStore):
+    def test_links_wikidata_qids_to_labels(self, label_store: LabelStore):
         enrichments = _make_enrichments({"Autechre": [("Warp Records", 23528)]})
         wikidata_client = MagicMock()
         wikidata_client.lookup_labels_by_discogs_ids.return_value = {
@@ -85,16 +106,16 @@ class TestPopulateLabelHierarchy:
         }
         wikidata_client.get_label_hierarchy.return_value = []
 
-        populate_label_hierarchy(store, enrichments, wikidata_client)
+        populate_label_hierarchy(label_store, enrichments, wikidata_client)
 
-        warp = store.get_label_by_name("Warp Records")
+        warp = _get_label_by_name(label_store._conn, "Warp Records")
         assert warp is not None
         assert warp["entity_id"] is not None
-        entity = store.get_entity_by_qid("Q1312934")
+        entity = _get_entity_by_qid(label_store._conn, "Q1312934")
         assert entity is not None
-        assert entity.name == "Warp Records"
+        assert entity["name"] == "Warp Records"
 
-    def test_populates_hierarchy(self, store: EntityStore):
+    def test_populates_hierarchy(self, label_store: LabelStore):
         enrichments = _make_enrichments(
             {
                 "Autechre": [("Warp Records", 23528)],
@@ -121,38 +142,36 @@ class TestPopulateLabelHierarchy:
             ),
         ]
 
-        populate_label_hierarchy(store, enrichments, wikidata_client)
+        populate_label_hierarchy(label_store, enrichments, wikidata_client)
 
         # Parent label should be auto-created
-        umg = store.get_label_by_name("Universal Music Group")
+        umg = _get_label_by_name(label_store._conn, "Universal Music Group")
         assert umg is not None
 
         # Hierarchy rows should exist
-        rows = store._conn.execute("SELECT COUNT(*) FROM label_hierarchy").fetchone()
+        rows = label_store._conn.execute("SELECT COUNT(*) FROM label_hierarchy").fetchone()
         assert rows[0] == 2
 
-    def test_skips_labels_without_discogs_id(self, store: EntityStore):
+    def test_skips_labels_without_discogs_id(self, label_store: LabelStore):
         enrichments = _make_enrichments({"Autechre": [("Self-Released", None)]})
         wikidata_client = MagicMock()
         wikidata_client.lookup_labels_by_discogs_ids.return_value = {}
         wikidata_client.get_label_hierarchy.return_value = []
 
-        populate_label_hierarchy(store, enrichments, wikidata_client)
+        populate_label_hierarchy(label_store, enrichments, wikidata_client)
 
-        # Label should still be created but no Discogs lookup attempted
-        label = store.get_label_by_name("Self-Released")
+        label = _get_label_by_name(label_store._conn, "Self-Released")
         assert label is not None
         assert label["discogs_label_id"] is None
-        # Should have called with empty list (no Discogs IDs to look up)
         wikidata_client.lookup_labels_by_discogs_ids.assert_called_once_with([])
 
-    def test_empty_enrichments(self, store: EntityStore):
+    def test_empty_enrichments(self, label_store: LabelStore):
         wikidata_client = MagicMock()
-        populate_label_hierarchy(store, {}, wikidata_client)
+        populate_label_hierarchy(label_store, {}, wikidata_client)
         wikidata_client.lookup_labels_by_discogs_ids.assert_not_called()
         wikidata_client.get_label_hierarchy.assert_not_called()
 
-    def test_deduplicates_labels_across_artists(self, store: EntityStore):
+    def test_deduplicates_labels_across_artists(self, label_store: LabelStore):
         enrichments = _make_enrichments(
             {
                 "Autechre": [("Warp Records", 23528)],
@@ -163,14 +182,14 @@ class TestPopulateLabelHierarchy:
         wikidata_client.lookup_labels_by_discogs_ids.return_value = {}
         wikidata_client.get_label_hierarchy.return_value = []
 
-        populate_label_hierarchy(store, enrichments, wikidata_client)
+        populate_label_hierarchy(label_store, enrichments, wikidata_client)
 
-        count = store._conn.execute(
+        count = label_store._conn.execute(
             "SELECT COUNT(*) FROM label WHERE name = 'Warp Records'"
         ).fetchone()[0]
         assert count == 1
 
-    def test_hierarchy_parent_from_outside_enrichments(self, store: EntityStore):
+    def test_hierarchy_parent_from_outside_enrichments(self, label_store: LabelStore):
         """Parent labels not in enrichments should still be created."""
         enrichments = _make_enrichments({"Autechre": [("Warp Records", 23528)]})
         wikidata_client = MagicMock()
@@ -186,15 +205,15 @@ class TestPopulateLabelHierarchy:
             ),
         ]
 
-        populate_label_hierarchy(store, enrichments, wikidata_client)
+        populate_label_hierarchy(label_store, enrichments, wikidata_client)
 
-        umg = store.get_label_by_name("Universal Music Group")
+        umg = _get_label_by_name(label_store._conn, "Universal Music Group")
         assert umg is not None
-        entity = store.get_entity_by_qid("Q21077")
+        entity = _get_entity_by_qid(label_store._conn, "Q21077")
         assert entity is not None
-        assert entity.entity_type == "label"
+        assert entity["entity_type"] == "label"
 
-    def test_returns_report(self, store: EntityStore):
+    def test_returns_report(self, label_store: LabelStore):
         enrichments = _make_enrichments(
             {
                 "Autechre": [("Warp Records", 23528)],
@@ -216,8 +235,8 @@ class TestPopulateLabelHierarchy:
             ),
         ]
 
-        report = populate_label_hierarchy(store, enrichments, wikidata_client)
+        report = populate_label_hierarchy(label_store, enrichments, wikidata_client)
 
         assert report.labels_created == 3
-        assert report.labels_matched == 2  # Warp + Sub Pop got QIDs
+        assert report.labels_matched == 2
         assert report.hierarchy_edges == 1
